chore: tidying up some of cursors files
Cursor leaves behind unused files, cleaning up and cleaning up the final prototype in general.
