--- apps/studio/components/interfaces/Account/NewAccessTokenSteps/index.ts
@@ -1,4 +1,0 @@
-export { default as Step1BasicInfo } from './Step1BasicInfo'
-export { default as Step2ResourceAccess } from './Step2ResourceAccess'
-export { default as Step3Permissions } from './Step3Permissions'
-export { default as Step4TokenGenerated } from './Step4TokenGenerated'
